Bug context: springboot 2.1.6 + webflux + tomcat(tested on 8.5.31/9.0.21/9.0.22) + gzip
Bug description: when the endpoint returns POJO/Collection/Mono/Mono<Collection> (not FLUX<POJO>), The unit test passed, but failed when call api by postman or curl

Cause:
in the response, contains both CONTENT_LENGTH and TRANSFER_ENCODING, so the response is not valid
in the response, the value of CONTENT_LENGTH is the one before compression, so the status is failed in chrome

Solution:
Clean the invalid headers using the NEW rule of http 1.1 protocol RFC7230, ignore the old rules of obsoleted RFC2616

Rule 1: A sender MUST NOT send a Content-Length header field in any message
that contains a Transfer-Encoding header field.
Rule 2: a server MUST NOT send Content-Length in a response unless its field-value equals the
decimal number of octets that would have been sent in the payload
body of a response.
for example, we should remove Content-Length if the compressed content length is unknown (Ref org.apache.coyote.CompressionConfig)